Fixed Dropdown Button Direction
Changed dropdownArea function to the original if-else statement so that the dropdown button swtiches direction when activated.

The new code I put in as suggested by the code review didn't work for some reason--I changed the suggested code a little bit because errors would pop up on VS Code, and the way I implemented it must not have been correct.