Draft #2: Saving progress before adding blockRendering
-Removed DirectionButtons from app, we will use forward button instead.

-ForwardButton.js is very similar to DirectionButtons, except we will set an interval to render post text/content slowly until it hits the end of a section.
After this commit, I will make it so that it will also change into also stopping once it hits an interactive section until it is marked "completed."

ForwardButton.js uses the useEffect hook to set/clear interval accordingly. ForwardButton will only render when the interval clears, since that is when the user can progress.

-Changed PostsList to contain all the states, since everything is mainly inside PostsList, no longer in App. (Might need to change since the Links on right/Progress bar will probably depend on visible sections as well.

-Added in a PlaceholderInteractive which I will make functional to stop the interval from progressing after this intermediate level commit.

-Clicking the button also hides its display until it is needed again.